pyproject.toml
[project]
name = "lios"
version = "0.1.0"
description = "Legal Intelligence Operating System for EU sustainability compliance"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "click>=8.1.7",
    "httpx>=0.27.0",
    "python-dateutil>=2.9.0",
    "rich>=13.7.0",
    "PyYAML>=6.0.1",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
llm = ["openai>=1.14.0"]
dev = ["pytest>=8.1.0", "pytest-asyncio>=0.23.0", "httpx>=0.27.0"]
data = [
    "httpx>=0.27.0",
    "lxml>=5.2.0",
    "beautifulsoup4>=4.12.0",
    "trafilatura>=1.9.0",
    "pypdf>=4.2.0",
    "feedparser>=6.0.11",
    "dateparser>=1.2.0",
    "rapidfuzz>=3.9.0",
    "sentence-transformers>=2.7.0",
    "rank-bm25>=0.2.2",
    "faiss-cpu>=1.8.0",
    "sqlalchemy>=2.0.30",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.3.0",
]

[project.scripts]
lios = "lios.cli.interface:cli"

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = { find = { include = ["lios*"] } }

[tool.pytest.ini_options]
asyncio_mode = "auto"
